+++ examples/demo-rollup/stf/src/tests/da_simulation.rs
@@ -2,21 +2,20 @@ use std::rc::Rc;
 
 use sov_bank::Bank;
 use sov_mock_da::MockDaSpec;
-use sov_modules_api::runtime::capabilities::Authenticator;
+use sov_modules_api::capabilities::RuntimeAuthenticator;
 use sov_modules_api::transaction::{Transaction, UnsignedTransaction};
-use sov_modules_api::{EncodeCall, PrivateKey, RawTx};
+use sov_modules_api::{EncodeCall, FullyBakedTx, PrivateKey, RawTx};
 use sov_test_utils::generators::bank::BankMessageGenerator;
 use sov_test_utils::generators::sequencer_registry::SequencerRegistryMessageGenerator;
 use sov_test_utils::generators::value_setter::{ValueSetterMessage, ValueSetterMessages};
 use sov_test_utils::{MessageGenerator, TestPrivateKey};
 
-use crate::authentication::ModAuth;
 use crate::runtime::Runtime;
 
 pub(crate) type S = sov_test_utils::TestSpec;
 type Da = MockDaSpec;
 
-pub fn simulate_da(admin: TestPrivateKey) -> Vec<RawTx> {
+pub fn simulate_da(admin: TestPrivateKey) -> Vec<FullyBakedTx> {
     let mut messages = Vec::default();
 
     let bank_generator = BankMessageGenerator::<S>::with_minter_and_transfer(admin.clone());
@@ -26,9 +25,7 @@ pub fn simulate_da(admin: TestPrivateKey) -> Vec<RawTx> {
         admin: Rc::new(admin),
         messages: vec![99, 33],
     }]);
-    messages.extend(
-        value_setter.create_default_raw_txs_without_gas_usage::<Runtime<S, Da>, ModAuth<S, Da>>(),
-    );
+    messages.extend(value_setter.create_default_encoded_txs_without_gas_usage::<Runtime<S, Da>>());
     let nonce_offset = messages.len() as u64;
     for mut msg in bank_messages {
         msg.nonce += nonce_offset;
@@ -38,16 +35,15 @@ pub fn simulate_da(admin: TestPrivateKey) -> Vec<RawTx> {
     messages
 }
 
-pub fn simulate_da_with_revert_msg(admin: TestPrivateKey) -> Vec<RawTx> {
+pub fn simulate_da_with_revert_msg(admin: TestPrivateKey) -> Vec<FullyBakedTx> {
     let mut messages = Vec::default();
     let bank_generator = BankMessageGenerator::<S>::create_invalid_transfer(admin);
-    let bank_txns =
-        bank_generator.create_default_raw_txs_without_gas_usage::<Runtime<S, Da>, ModAuth<S, Da>>();
+    let bank_txns = bank_generator.create_default_encoded_txs_without_gas_usage::<Runtime<S, Da>>();
     messages.extend(bank_txns);
     messages
 }
 
-pub fn simulate_da_with_bad_sig(key: TestPrivateKey) -> Vec<RawTx> {
+pub fn simulate_da_with_bad_sig(key: TestPrivateKey) -> Vec<FullyBakedTx> {
     let bank_generator: BankMessageGenerator<S> = BankMessageGenerator::with_minter(key.clone());
     let create_token_message = bank_generator.create_default_messages().remove(0);
     let tx = Transaction::<S>::new_with_details(
@@ -63,7 +59,7 @@ pub fn simulate_da_with_bad_sig(key: TestPrivateKey) -> Vec<RawTx> {
     vec![encode_with_auth(tx)]
 }
 
-pub fn simulate_da_with_bad_nonce(key: TestPrivateKey) -> Vec<RawTx> {
+pub fn simulate_da_with_bad_nonce(key: TestPrivateKey) -> Vec<FullyBakedTx> {
     let bank_generator: BankMessageGenerator<S> = BankMessageGenerator::with_minter(key);
     let mut create_token_message = bank_generator.create_default_messages().remove(0);
     // Overwrite the nonce with the maximum value
@@ -72,7 +68,7 @@ pub fn simulate_da_with_bad_nonce(key: TestPrivateKey) -> Vec<RawTx> {
     vec![encode_with_auth(tx)]
 }
 
-pub fn simulate_da_with_bad_serialization(key: TestPrivateKey) -> Vec<RawTx> {
+pub fn simulate_da_with_bad_serialization(key: TestPrivateKey) -> Vec<FullyBakedTx> {
     let bank_generator: BankMessageGenerator<S> = BankMessageGenerator::with_minter(key);
     let create_token_message = bank_generator.create_default_messages().remove(0);
     let tx = Transaction::<S>::new_signed_tx(
@@ -87,9 +83,9 @@ pub fn simulate_da_with_bad_serialization(key: TestPrivateKey) -> Vec<RawTx> {
     vec![encode_with_auth(tx)]
 }
 
-fn encode_with_auth(tx: Transaction<S>) -> RawTx {
-    let tx_bytes = borsh::to_vec(&tx).unwrap();
-    ModAuth::<S, Da>::encode(tx_bytes).unwrap()
+fn encode_with_auth(tx: Transaction<S>) -> FullyBakedTx {
+    let tx_bytes = RawTx::new(borsh::to_vec(&tx).unwrap());
+    Runtime::<S, MockDaSpec>::encode_with_standard_auth(tx_bytes)
 }
 
 pub fn simulate_da_with_incorrect_direct_registration_msg(admin: TestPrivateKey) -> RawTx {